Move server.py into its own module
